Update micromamba section
This section is now improved and more similar to what had been
shown before with conda.

There was also a bug in which salmon,
fastqc and multiqc couldn't be found, and then fastqc would
fail due to a missing perl module. This has been fixed by
fixing the order of the channels and setting an ENV at the
Dockerfile.

Signed-off-by: Marcel Ribeiro-Dantas <mribeirodantas@seqera.io>