Have a single env.yml in the training material
Historically, the env.yml file content in the training has been
different from the one already existing in the file. As both used
to work, both were maintained this way. However, currently only the
one written in the material (not in the file) work, so this commit
adds consistency by having all env.yml equal to the one first mentioned
in the training material, which works.

Signed-off-by: Marcel Ribeiro-Dantas <mribeirodantas@seqera.io>