Update final_main.nf to use multiple containers
This commit updates the final_main.nf to follow the best
practices, i.e. have a light container image for every
process. The limitation for now is the latest process that has
a tricky container image with many R dependencies and we'll let
the bulky crg container image for now.

Signed-off-by: Marcel Ribeiro-Dantas <mribeirodantas@seqera.io>