Add -p to mkdir in script blocks
Depending on how a pipeline is stopped (e.g. CTRL-C), Nextflow
may not have the opportunity to finish things properly, which
will make it attempt to run the task again but on the same path.
If you're trying to use `mkdir` and the folder is already there,
there will be an error and the task will fail. The `-p` will make
mkdir not complain if the folder is already there.

Signed-off-by: Marcel Ribeiro-Dantas <mribeirodantas@seqera.io>