Moved intro to variable inputs and params back to Hello World
After reconsidering the review comments, realized that it is ok and actually better to introduce variable inputs and params in Hello World. Getting them out of the way before Channels is actually helpful. The trick is to not get bogged down in details about value versus queue channels. The rest is a bit streamlined as a results. Also made Operators intro more explicit.

Also got started on some cleanup of the code files