Big wave of cleanup
Moved non-maintained materials to 'other' directory, commented out the future courses that aren't ready yet, and updated index pages and navs accordingly.

Also turned off alt languages because the translations are too far out of date, to the point of being misleading. We'll bring back translations when we've had a chance to refresh them. (They will still be accessible via the v2.0 of the portal)