Add Survey and Next Steps pages (#524)
Add a Survey page at the end of the training to collect feedback, then a final one with links to further resources. The pages are set up to encourage learners to complete the survey in order to access the next steps, but the Next Steps page is accessible directly from the left hand menu. 

Just doing this for Hello Nextflow for now but if it works well we'll clone this for all other trainings.